[VSC-7] Change moc.js functions for the LSP
Problem: In the compiler, we changed the `parseMotokoTyped` function for
the LSP that allows caching for better performance. This needs to be
exposed by node-motoko. Moreover, `parseMotokoWithDeps` was added which
parses the file and returns immediate dependencies.

Solution: Regenerate based on a local compiler build. Change `file.ts`
and `index.ts` to account for these functions.